Backend folder restructure and code quality improvised
