+++ .gitignore
@@ -1,3 +1,4 @@
-.env
-package-lock.json
 node_modules
+**/node_modules
+
+**/.env
